Move example to the top-level dir (rpl-border-router) and tidy-up webserver configuration
Previously, we were using the example's Makefile to configure whether we want a web server built into the BR. We had 3 options:

* No web server
* Internal web server
* External web server (from apps/)

The last option is no longer evailable. This commit removes support for this option and it also simplifies the situation: We now merely use a CPP macro to enable/disable the web server.